requirements-extra.txt
# Additional optional features
llama-index-agent-openai==0.4.2
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.14
llama-index-multi-modal-llms-openai==0.4.2
openai==1.60.1